+++ Server/Source.cpp
@@ -6,27 +6,27 @@ using namespace std::chrono;
 
 int main()
 {
-	auto start = high_resolution_clock::now();
+    auto start = high_resolution_clock::now();
 
-	if (Network::Initialize())
-	{
-		MyServer server;
+    if (Network::Initialize())
+    {
+        MyServer server;
 
-		if (server.Initialize(IPEndpoint("0.0.0.0", 6112)))
-		{
-			while (true)
-			{
-				server.Frame();
-				float tim = duration_cast<milliseconds>(high_resolution_clock::now() - start).count();
-				if (tim > 20)
-				{
-					start = high_resolution_clock::now();
-					server.SendPackets();
-				}
-			}
-		}
-	}
-	Network::Shutdown();
-	//system("pause");
-	return 0;
+        if (server.Initialize(IPEndpoint("0.0.0.0", 6112)))
+        {
+            while (true)
+            {
+                server.Frame();
+                uint64_t tim = duration_cast<milliseconds>(high_resolution_clock::now() - start).count();
+                if (tim > 20)
+                {
+                    start = high_resolution_clock::now();
+                    server.SendPackets();
+                }
+            }
+        }
+    }
+    Network::Shutdown();
+    //system("pause");
+    return 0;
 }
